Upadte public APIs and internal data structure.
CollisionDetection2d APIs are simplified.
In ver 1.0.x, user need to feed 4 parameters to construct container.
(Which are extrema, intersects, getBoundingBox, and boundary)
Now, it is reduced to 3 (range, boundingBox, isCollided).

Also, internal data structure is changed.
In ver 1.0.x, array is used in order to reduce time of put/get items.
Now, hand-written tree structure is adopted and this makes the code of
collision-detection function simplified.

The functions related to Index previously provided in
CollisionDetection2d/Internal/CollisionDetection2d
is moved to CollisionDetection2d/Internal/Index2d.

Examples and benchmark is temporary deleted since they are not fit to
this version.